update cost model
